+++ text-ui-test/EXPECTED.TXT
@@ -1,7 +1,83 @@
-Hello from
- ____        _        
-|  _ \ _   _| | _____ 
-| | | | | | | |/ / _ \
-| |_| | |_| |   <  __/
-|____/ \__,_|_|\_\___|
-
+	_____________________________________________________________________
+	 Awakening.... 
+	 Hi! I'm Faiz!
+	 What can I do for you?
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Exceed Charge.... 
+	 OOPS!!! Unable to mark invalid task as done.
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Here are the tasks in your list:
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Exceed Charge.... 
+	 OOPS!!! The description of a todo task cannot be empty.
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Exceed Charge.... 
+	 OOPS!!! The description of a deadline task cannot be empty.
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Exceed Charge.... 
+	 OOPS!!! The description of an event task cannot be empty.
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Got it. I've added this task: 
+	   [T][ ] borrow book
+	 Now you have 1 task in the list.
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Got it. I've added this task: 
+	   [D][ ] return book (by: saturday)
+	 Now you have 2 tasks in the list.
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Got it. I've added this task: 
+	   [E][ ] meet friends (from: 5pm tdy to: 9pm tdy)
+	 Now you have 3 tasks in the list.
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Exceed Charge.... 
+	 OOPS!!! Unable to mark invalid task as done.
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Nice! I've marked this task as done:
+	   [D][X] return book (by: saturday)
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Here are the tasks in your list:
+	 1.[T][ ] borrow book
+	 2.[D][X] return book (by: saturday)
+	 3.[E][ ] meet friends (from: 5pm tdy to: 9pm tdy)
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Exceed Charge.... 
+	 OOPS!!! Unable to mark invalid task as undone.
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 OK, I've marked this task as not done yet:
+	   [D][ ] return book (by: saturday)
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Here are the tasks in your list:
+	 1.[T][ ] borrow book
+	 2.[D][ ] return book (by: saturday)
+	 3.[E][ ] meet friends (from: 5pm tdy to: 9pm tdy)
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 ERROR.... 
+	 A thousand apologies, but I don't know what that means.
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Exceed Charge.... 
+	 OOPS!!! The index of a task to be marked as done cannot be empty.
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Exceed Charge.... 
+	 OOPS!!! The index of a task to be marked as undone cannot be empty.
+	_____________________________________________________________________
+	_____________________________________________________________________
+	 Reformation.... 
+	 Bye! Hope to see you again soon!
+	_____________________________________________________________________
