+++ text-ui-test/input.txt
@@ -1,0 +1,18 @@
+mark 0
+list
+todo
+deadline
+event
+todo borrow book
+deadline return book /by saturday
+event meet friends /from 5pm tdy /to 9pm tdy
+mark 0
+mark 2
+list
+unmark 10
+unmark 2
+list
+blah
+mark
+unmark
+bye
